ocms/requirements.txt
Django>=5.0,<6.1
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
requests
